Update the Shiny app flow and refactor the server code
A single reactive block of code watches relevant inputs and executes ICY's
main individual choosing process. This block checks for required inputs
and validates them.

The renderTable and  renderPlot elements now watch the output of
said reactive block of code, and  use the output to generate their results.

An event observer watches for changes in input file parameters and
handles the reading of the studbook information. These input data
are stored in "state" as reactive values, that - if changed - make the
ICY chooser reactive block to update, and pass results to the
renderTable and renderPlot appropriately.